issue #18 solved
