+++ reports/metrics.json
@@ -1,0 +1,26 @@
+{
+    "Ensemble": {
+        "rmse": 12.2068,
+        "mae": 12.2068,
+        "mape": 12.4559,
+        "r2": NaN
+    },
+    "ARIMA": {
+        "rmse": 33.1496,
+        "mae": 33.1496,
+        "mape": 33.8261,
+        "r2": NaN
+    },
+    "XGBoost": {
+        "rmse": 1.2721,
+        "mae": 1.2721,
+        "mape": 1.2981,
+        "r2": NaN
+    },
+    "LSTM": {
+        "rmse": 98.0,
+        "mae": 98.0,
+        "mape": 100.0,
+        "r2": NaN
+    }
+}
