+++ src/models/__init__.py
@@ -1,3 +1,4 @@
 from .lstm_model import LSTMModel, SlidingWindowDataset
+from .ensemble_model import EnsembleForecaster
 from .xgb_model import train_xgb
 from .arima_model import train_arima
